requirements.txt
Django==4.1.7
djangorestframework==3.14.0
python-dotenv==1.0.1
Requests==2.31.0
google-cloud-storage==2.13.0
Pillow==10.2.0
psycopg2-binary==2.9.9
beautifulsoup4==4.12.3
python-Levenshtein>=0.21.1
fuzzywuzzy>=0.18.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.108.0
google-generativeai==0.3.2
tenacity==8.2.3
pandas==2.1.4
pyarrow==14.0.0
sqlalchemy==2.0.30
tqdm==4.67.0
django-filter==23.5
gspread==5.12.0
whitenoise==6.6.0
django-cors-headers==4.3.1
drf_yasg==1.21.7
adrf==0.1.6 